requirements.txt
streamlit
tensorflow-cpu==2.13.0
numpy
matplotlib
pillow
opencv-python-headless
scikit-image
lime